pyproject.toml
[project]
name = "myxo"
version = "0.1.0"
description = "A Fold-native autonomous AI creature"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "openai>=1.50.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pymupdf>=1.24.0",
]

[tool.setuptools.packages.find]
include = ["myxo*"]
